requirements.txt
python-telegram-bot
argparse
python-Wappalyzer
setuptools>=58.1.0
spyse-python>=2.2.3
starlette>=0.14.2
chardet>=3.0.4
idna>=2.9
urllib3>=1.25.8
PySocks>=1.7.1
certifi>=2020.4.5.1
Jinja2>=3.0.0
pyyaml>=5.3.1
cryptography>=2.8
cffi>=1.14.0
markupsafe>=2.0.1
pyopenssl>=21.0.0
charset_normalizer~=2.0.0
requests>=2.25.0
requests_ntlm>=1.1.0
ntlm_auth>=1.5.0
pyparsing>=3.0.6
python-magic>=0.4.24